consolidate platforms with non-delegating impl (#350)
create directMain source set as base for all non-delegating platforms - platforms that use kotlin-logging to 
write the logs with native kotlin functions and not via other libs or platform constructs.

This change the js/linux/etc' platforms. Some duplication was removed. And the way it is now configured changed.
`KotlinLoggingConfiguration` now contains:
- logLevel
- formatter
- appender

Those interfaces changed. Logger now calls the appender that is responsible to call the formatter - unlike previously where it called the formatter and passed the message to the appender.
Appender don't need to check if log is enabled, and has a single method:
`fun log(loggingEvent: KLoggingEvent)` (which contains also payload).
Formatter has a single method:
`fun formatMessage(loggingEvent: KLoggingEvent): String` which returns a formatted message.


* old upper case names in js are now deprecated and delegating to those.